// src/app/types.ts
export type InstitutionProps = {
  name: string;
  code: string;
  type: string;
};